Fix printing issue bug
Fixed an issue with printing and incidentally probably fixed a bigger bug with options not being passed correctly. I think it wasn't picked up in testing as the protoessentials in the test case of extraspecial group of order 3^3 returns the same protoessentials regardless of OpTriv and pPerfect so AllFusionSystems worked as expected and produced two different results coincidentally.