Update direct-product.m
Added a bit of robustness for working with the DirectProductGroup type, unsure if it is useful yet.
Otherwise looks to be working, the few cases I've tested have given expected results but more testing needed of course.
Potential issue is when you pass copies of a fusion system i.e. call FusionDirectProduct(F,F)