Script to split files
Still a long way to go on this issue.
split-file.py is a script that when run will split an Allpton file into the correct folder structure, of course there remains the ones that are ad-hoc or not in a single file.

Began to write a converter script to run in MAGMA that will take one of these FS_i files and output a nicer recorded format. The idea being that currently to browse the database we have to actually load every single fusion system, here we just really want data with no execution.

Ideally we would have the following
SmallFusionSystem(S, i) which will return the fusion system over S with index i
IdentifyFusionSystem(F) which will determine if F is already in our database, to speed this up we might want to note the number of F`essentials in the database records
AddSmallFusionSystem(F), given a fusion system F, it is not currently in the database then add it.

Priority is to get a decent converter script for the new record format, still got some little irritating bugs to iron out with regards to MAGMA interacting with cmd